Spell: Fix implementation of resistance percentage mods wrongly altering unitfields.
Credit to cmangos.

https://github.com/cmangos/mangos-classic/commit/9439705c8213841b288e63075301b406cb6f6fc1

+++ src/game/Objects/Unit.cpp
@@ -7497,7 +7497,7 @@ float Unit::GetTotalStatValue(Stats stat) const
     return value;
 }
 
-int32 Unit::GetTotalResistanceValue(SpellSchools school) const
+float Unit::GetTotalResistanceValue(SpellSchools school) const
 {
     UnitMods unitMod = UnitMods(UNIT_MOD_RESISTANCE_START + school);
 
@@ -7524,7 +7524,7 @@ int32 Unit::GetTotalResistanceValue(SpellSchools school) const
         value = 0;
 #endif
 
-    return int32(value);
+    return value;
 }
 
 float Unit::GetTotalAuraModValue(UnitMods unitMod) const
